Adding the new sidebar on the right for Farms.
We redesigned the navigation to make it more intuitive. In this commit,
I'm implementing the right sidebar to navigate between farms. So, when
the user navigate by using the left sidebar they only navigate inside
single farm. If they want to jump to another farm, they will click the
farm menu on the right sidebar, then the entire Tania data will show
only the data for that particular farm.