Moving some default data from Fixtures to Migrations.
Since `doctrine:fixtures:load` purges the database, so I move the
default data seeding process to doctrine migration. It will also avoid
the database to be filled with a test user account.

`doctrine:fixtures:load` is called only in the testing environment. I'm
also adjusting the installation documentation for this.